Create test input file generators
Test input files are written in Python to make them simple and quick to
write; but I do not want to have any Python get sent to the build. I may
even have CMake run a Python file that finds all the other Python files
in the test/testInput directory that have output file creators.

Because the Python files run extremely fast, and because the file edits
may not make it to CMake, I may want the build system to call Python
every time that is called. I am not sure all build systems would allow
for arbitrary program calls.